+++ accounts-add-service-server.js
@@ -3,6 +3,70 @@
 "use strict";
 var mergeUserErrorReason = AccountsAddService._mergeUserErrorReason;
 
+AccountsAddService.databaseMigrationEnabled = true;
+
+AccountsAddService._migrationControl = 
+  new Mongo.Collection("AccountsAddService._migrationControl");
+
+// returns false if the migration isn't run, or the number of users updated
+// if it is.
+AccountsAddService._migrateDatabase = function () {
+  if (! AccountsAddService.databaseMigrationEnabled) {
+    return false;
+  }
+  var addHasLoggedIn = 
+    AccountsAddService._migrationControl.findOne('addHasLoggedIn');
+  if (addHasLoggedIn && addHasLoggedIn.startedAt) {
+    return false;
+  }
+  if (! addHasLoggedIn) {
+    try {
+      AccountsAddService._migrationControl.insert({ _id: 'addHasLoggedIn' });
+    } catch (err) {
+      // Ignore duplicate key error thrown if already id already exists due to
+      // concurrent insertion attempts
+      if (! (err.name === 'MongoError' && err.message.indexOf('E11000'))) {
+        throw err;
+      }
+    }
+  }
+  var numAffected = AccountsAddService._migrationControl.update({
+    _id: 'addHasLoggedIn',
+    startedAt: { $exists: false }
+   }, {
+    $set: {
+      startedAt: new Date(),
+    }
+  });
+  // Only one server will return numAffected !== 0. When numAffect === 0,
+  // The migration was already started (and possibly finished)
+  if (numAffected === 0) {
+    return false;
+  }
+  
+  var selector = { 
+    hasLoggedIn: { $exists: false }, 
+    'services.resume': { $exists: true }
+  };
+  var numUsersUpdated = Meteor.users.update(selector, {
+    $set: { hasLoggedIn: true }
+  });
+  if (numUsersUpdated > 0) {
+    console.log('brettle:accounts-add-service set hasLoggedIn = true for ' +
+      numUsersUpdated + ' existing user(s).');  
+  }
+  AccountsAddService._migrationControl.update({
+    _id: 'addHasLoggedIn',
+   }, {
+    $set: {
+      finishedAt: new Date(),
+    }
+  });
+  return numUsersUpdated;
+};
+
+Meteor.startup(AccountsAddService._migrateDatabase);
+
 // The first time a user logs in, we set his hasLoggedIn property so that
 // we don't accidentally merge his account if his "resume" service is removed.
 Accounts.onLogin(function (attemptInfo) {